Calculates Great Circle distance between points
Calculates the Great Circle distance in miles between two points: a
point of interest and a landmark (defined here as hospitals,
universities, and an employer among the 25 largest in the same industry
in the same state).

Requires National Establishment Time Series database, or can be adapted
to any data set with lat/lon coordinate information.